chore: Move WithUsageDataPreviewPanel to the MetricVizPanel folder
